+++ src/Model/Command/ItemPropertySetup.php
@@ -2,6 +2,8 @@
 
 namespace Lmc\Matej\Model\Command;
 
+use Lmc\Matej\Model\Assertion;
+
 /**
  * Command to add or delete item property in the database.
  */
@@ -21,10 +23,7 @@ class ItemPropertySetup extends AbstractCommand
 
     private function __construct(string $propertyName, string $propertyType)
     {
-        // TODO: assert propertyName format
-        // TODO: assert propertyType is one of PROPERTY_TYPE_*
-
-        $this->propertyName = $propertyName;
+        $this->setPropertyName($propertyName);
         $this->propertyType = $propertyType;
     }
 
@@ -56,6 +55,13 @@ class ItemPropertySetup extends AbstractCommand
     public static function set(string $propertyName): self
     {
         return new static($propertyName, self::PROPERTY_TYPE_SET);
+    }
+
+    protected function setPropertyName(string $propertyName): void
+    {
+        Assertion::typeIdentifier($propertyName);
+
+        $this->propertyName = $propertyName;
     }
 
     protected function getCommandType(): string